[Refactor] Split storage primitive value types
