+++ be/src/exec/olap_scan_node.cpp
@@ -54,7 +54,8 @@
 #include "runtime/exec_env.h"
 #include "runtime/global_dict/fragment_dict_state.h"
 #include "storage/chunk_helper.h"
-#include "storage/olap_common.h"
+#include "storage/primitive/storage_ids.h"
+#include "storage/primitive/storage_version.h"
 #include "storage/rowset/rowset.h"
 #include "storage/storage_engine.h"
 #include "storage/tablet.h"
