+++ be/src/tools/meta_tool.cpp
@@ -50,6 +50,7 @@
 #include "base/path/path_util.h"
 #include "column/chunk_factory.h"
 #include "column/datum_convert.h"
+#include "common/column_id.h"
 #include "common/config_exec_fwd.h"
 #include "common/config_storage_fwd.h"
 #include "common/configbase.h"
@@ -78,9 +79,9 @@
 #include "storage/key_coder.h"
 #include "storage/lake/tablet_manager.h"
 #include "storage/lake/vacuum.h"
-#include "storage/olap_common.h"
 #include "storage/olap_define.h"
 #include "storage/primary_key_dump.h"
+#include "storage/primitive/storage_stats.h"
 #include "storage/rowset/binary_plain_page.h"
 #include "storage/rowset/column_iterator.h"
 #include "storage/rowset/column_reader.h"
